カスタムエージェントの設定画面を改善
