+++ .eslintrc.cjs
@@ -14,6 +14,7 @@ module.exports = {
   rules: {
     '@typescript-eslint/explicit-function-return-type': 'off',
     '@typescript-eslint/no-explicit-any': 'off',
+    'react/prop-types': 'off',
     'no-control-regex': 0
   }
 }
